pin are very limited, having them in byte is more than enouth. add a repeat transmit to init the rf433

+++ Transmitter/Transmitter.ino
@@ -10,7 +10,7 @@ get 32 bits values.
 
 
 // Transmitter params
-int txPin = 12;
+byte txPin = 12;
 
 Transmitter tx = Transmitter(txPin);
